- Handled initialisation in the RudderSdkFlutterPlugin instead of the Application file.
