+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.rudderstack.rudder_sdk_flutter">
-<application android:name="com.rudderstack.sdk.flutter.RudderSdkFlutterApplication"/>
+    package="com.rudderstack.rudder_sdk_flutter">
+
 </manifest>
